requirements.txt
pandas~=1.3.5
numpy~=1.21.2
matplotlib~=3.5.1
scipy
scikit-learn~=1.0.2
pytorch
spacy~=3.2.1